backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
langchain
langchain-google-genai
langgraph
langsmith
pydantic
